Ch. 17: renumber listings to eliminate `no-listing` listings
