+++ listings/ch17-async-await/listing-17-19/Cargo.toml
@@ -4,6 +4,3 @@ version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-trpl = { path = "../../../packages/trpl" }
