+++ listings/ch17-async-await/listing-17-21/src/main.rs
@@ -1,15 +1,11 @@
-use std::{
-    future::Future,
-    pin::{pin, Pin},
-    time::Duration,
-};
+use std::{future::Future, time::Duration};
 
 fn main() {
     trpl::block_on(async {
         let (tx, mut rx) = trpl::channel();
 
         let tx1 = tx.clone();
-        let tx1_fut = pin!(async move {
+        let tx1_fut = async move {
             let vals = vec![
                 String::from("hi"),
                 String::from("from"),
@@ -21,15 +17,15 @@ fn main() {
                 tx1.send(val).unwrap();
                 trpl::sleep(Duration::from_secs(1)).await;
             }
-        });
+        };
 
-        let rx_fut = pin!(async {
+        let rx_fut = async {
             while let Some(value) = rx.recv().await {
                 println!("received '{value}'");
             }
-        });
+        };
 
-        let tx_fut = pin!(async move {
+        let tx_fut = async move {
             let vals = vec![
                 String::from("more"),
                 String::from("messages"),
@@ -41,11 +37,11 @@ fn main() {
                 tx.send(val).unwrap();
                 trpl::sleep(Duration::from_secs(1)).await;
             }
-        });
+        };
 
         // ANCHOR: here
-        let futures: Vec<Pin<Box<dyn Future<Output = ()>>>> =
-            vec![Box::pin(tx1_fut), Box::pin(rx_fut), Box::pin(tx_fut)];
+        let futures: Vec<Box<dyn Future<Output = ()>>> =
+            vec![Box::new(tx1_fut), Box::new(rx_fut), Box::new(tx_fut)];
         // ANCHOR_END: here
 
         trpl::join_all(futures).await;
