+++ listings/ch17-async-await/listing-17-24/src/main.rs
@@ -1,21 +1,62 @@
-use std::time::Duration;
+use std::{
+    future::Future,
+    pin::{pin, Pin},
+    time::Duration,
+};
 
 fn main() {
     trpl::block_on(async {
+        let (tx, mut rx) = trpl::channel();
+
+        let tx1 = tx.clone();
         // ANCHOR: here
-        let slow = async {
-            println!("'slow' started.");
-            trpl::sleep(Duration::from_millis(100)).await;
-            println!("'slow' finished.");
-        };
+        let tx1_fut = pin!(async move {
+            // snip...
+            // ANCHOR_END: here
+            let vals = vec![
+                String::from("hi"),
+                String::from("from"),
+                String::from("the"),
+                String::from("future"),
+            ];
 
-        let fast = async {
-            println!("'fast' started.");
-            trpl::sleep(Duration::from_millis(50)).await;
-            println!("'fast' finished.");
-        };
+            for val in vals {
+                tx1.send(val).unwrap();
+                trpl::sleep(Duration::from_secs(1)).await;
+            }
+            // ANCHOR: here
+        });
 
-        trpl::race(slow, fast).await;
+        let rx_fut = pin!(async {
+            // snip...
+            // ANCHOR_END: here
+            while let Some(value) = rx.recv().await {
+                println!("received '{value}'");
+            }
+            // ANCHOR: here
+        });
+
+        let tx_fut = pin!(async move {
+            // snip...
+            // ANCHOR_END: here
+            let vals = vec![
+                String::from("more"),
+                String::from("messages"),
+                String::from("for"),
+                String::from("you"),
+            ];
+
+            for val in vals {
+                tx.send(val).unwrap();
+                trpl::sleep(Duration::from_secs(1)).await;
+            }
+            // ANCHOR: here
+        });
+
+        let futures: Vec<Pin<&mut dyn Future<Output = ()>>> =
+            vec![tx1_fut, rx_fut, tx_fut];
         // ANCHOR_END: here
+
+        trpl::join_all(futures).await;
     });
 }
