+++ listings/ch17-async-await/listing-17-26/src/main.rs
@@ -1,35 +1,21 @@
-use std::{thread, time::Duration};
+use std::time::Duration;
 
 fn main() {
     trpl::block_on(async {
-        // ANCHOR: slow-futures
-        let a = async {
-            println!("'a' started.");
-            slow("a", 300);
-            slow("a", 100);
-            slow("a", 200);
-            slow("a", 900);
-            trpl::sleep(Duration::from_millis(50)).await;
-            println!("'a' finished.");
+        // ANCHOR: here
+        let slow = async {
+            println!("'slow' started.");
+            trpl::sleep(Duration::from_millis(100)).await;
+            println!("'slow' finished.");
         };
 
-        let b = async {
-            println!("'b' started.");
-            slow("b", 750);
-            slow("b", 100);
-            slow("b", 150);
-            slow("b", 350);
-            slow("b", 150);
+        let fast = async {
+            println!("'fast' started.");
             trpl::sleep(Duration::from_millis(50)).await;
-            println!("'b' finished.");
+            println!("'fast' finished.");
         };
 
-        trpl::race(a, b).await;
-        // ANCHOR_END: slow-futures
+        trpl::race(slow, fast).await;
+        // ANCHOR_END: here
     });
-}
-
-fn slow(name: &str, ms: u64) {
-    thread::sleep(Duration::from_millis(ms));
-    println!("'{name}' ran for {ms}ms");
 }
