+++ listings/ch17-async-await/listing-17-29/src/main.rs
@@ -1,34 +1,44 @@
-use std::time::{Duration, Instant};
+use std::{thread, time::Duration};
 
 fn main() {
     trpl::block_on(async {
         // ANCHOR: here
-        let one_ns = Duration::from_nanos(1);
-        let start = Instant::now();
-        async {
-            for _ in 1..1000 {
-                trpl::sleep(one_ns).await;
-            }
-        }
-        .await;
-        let time = Instant::now() - start;
-        println!(
-            "'sleep' version finished after {} seconds.",
-            time.as_secs_f32()
-        );
+        let one_ms = Duration::from_millis(1);
 
-        let start = Instant::now();
-        async {
-            for _ in 1..1000 {
-                trpl::yield_now().await;
-            }
-        }
-        .await;
-        let time = Instant::now() - start;
-        println!(
-            "'yield' version finished after {} seconds.",
-            time.as_secs_f32()
-        );
+        let a = async {
+            println!("'a' started.");
+            slow("a", 300);
+            trpl::sleep(one_ms).await;
+            slow("a", 100);
+            trpl::sleep(one_ms).await;
+            slow("a", 200);
+            trpl::sleep(one_ms).await;
+            slow("a", 900);
+            trpl::sleep(one_ms).await;
+            println!("'a' finished.");
+        };
+
+        let b = async {
+            println!("'b' started.");
+            slow("b", 750);
+            trpl::sleep(one_ms).await;
+            slow("b", 100);
+            trpl::sleep(one_ms).await;
+            slow("b", 150);
+            trpl::sleep(one_ms).await;
+            slow("b", 350);
+            trpl::sleep(one_ms).await;
+            slow("b", 150);
+            trpl::sleep(one_ms).await;
+            println!("'b' finished.");
+        };
         // ANCHOR_END: here
+
+        trpl::race(a, b).await;
     });
+}
+
+fn slow(name: &str, ms: u64) {
+    thread::sleep(Duration::from_millis(ms));
+    println!("'{name}' ran for {ms}ms");
 }
