+++ listings/ch17-async-await/listing-17-31/src/main.rs
@@ -1,0 +1,34 @@
+use std::time::{Duration, Instant};
+
+fn main() {
+    trpl::block_on(async {
+        // ANCHOR: here
+        let one_ns = Duration::from_nanos(1);
+        let start = Instant::now();
+        async {
+            for _ in 1..1000 {
+                trpl::sleep(one_ns).await;
+            }
+        }
+        .await;
+        let time = Instant::now() - start;
+        println!(
+            "'sleep' version finished after {} seconds.",
+            time.as_secs_f32()
+        );
+
+        let start = Instant::now();
+        async {
+            for _ in 1..1000 {
+                trpl::yield_now().await;
+            }
+        }
+        .await;
+        let time = Instant::now() - start;
+        println!(
+            "'yield' version finished after {} seconds.",
+            time.as_secs_f32()
+        );
+        // ANCHOR_END: here
+    });
+}
