--- listings/ch17-async-await/no-listing-borrow-mutable-vec/src/main.rs
@@ -1,18 +1,0 @@
-fn main() {
-    trpl::block_on({
-        // ANCHOR: here
-        async {
-            let mut strings = vec![];
-
-            let a = trpl::read_to_string("test-data/hello.txt").await.unwrap();
-            strings.push(a.trim());
-
-            let b = trpl::read_to_string("test-data/world.txt").await.unwrap();
-            strings.push(b.trim());
-
-            let combined = strings.join(" ");
-            println!("{combined}");
-        }
-        // ANCHOR_END: here
-    });
-}
