--- listings/ch17-async-await/no-listing-type-mismatch/src/main.rs
@@ -1,7 +1,0 @@
-fn main() {
-    // ANCHOR: here
-    let a = 1;
-    let b = "Hello";
-    let vals = vec![a, b];
-    // ANCHOR_END: here
-}
