+++ src/ch17-04-more-ways-of-combining-futures.md
@@ -6,16 +6,16 @@ we move on. Sometimes, though, we only need *some* future from a set to finish
 before we move on—kind of like racing one future against another. This operation
 is often named `race` for exactly that reason.
 
-In Listing 17-24, we use `race` to run two futures, `slow` and `fast`, against
+In Listing 17-26, we use `race` to run two futures, `slow` and `fast`, against
 each other. Each one prints a message when it starts running, pauses for some
 amount of time by calling and awaiting `sleep`, and then prints another message
 when it finishes. Then we pass both to `trpl::race` and wait for one of them to
 finish. (The outcome here won’t be too surprising: `fast` wins!)
 
-<Listing number="17-24" caption="Using `race` to get the result of whichever future finishes first" file-name="src/main.rs">
+<Listing number="17-26" caption="Using `race` to get the result of whichever future finishes first" file-name="src/main.rs">
 
 ```rust
-{{#rustdoc_include ../listings/ch17-async-await/listing-17-24/src/main.rs:here}}
+{{#rustdoc_include ../listings/ch17-async-await/listing-17-26/src/main.rs:here}}
 ```
 
 </Listing>
@@ -59,22 +59,22 @@ function “slow” will just take a number of milliseconds to run, and sleep t
 thread for that long. This is intentionally not an async function, because the
 idea is to represent work that is *not* async.
 
-<Listing number="17-25" caption="Using `thread::sleep` to simulate slow operations" file-name="src/main.rs">
+<Listing number="17-27" caption="Using `thread::sleep` to simulate slow operations" file-name="src/main.rs">
 
 ```rust
-{{#rustdoc_include ../listings/ch17-async-await/listing-17-25/src/main.rs:slow}}
+{{#rustdoc_include ../listings/ch17-async-await/listing-17-27/src/main.rs:slow}}
 ```
 
 </Listing>
 
-In Listing 17-26, we use `slow` to emulate doing this kind of CPU-bound work in
+In Listing 17-28, we use `slow` to emulate doing this kind of CPU-bound work in
 a pair of futures. To begin, each future only hands control back to the runtime
 *after* carrying out a bunch of slow operations.
 
-<Listing number="17-26" caption="Using `thread::sleep` to simulate slow operations" file-name="src/main.rs">
+<Listing number="17-28" caption="Using `thread::sleep` to simulate slow operations" file-name="src/main.rs">
 
 ```rust
-{{#rustdoc_include ../listings/ch17-async-await/listing-17-26/src/main.rs:slow-futures}}
+{{#rustdoc_include ../listings/ch17-async-await/listing-17-28/src/main.rs:slow-futures}}
 ```
 
 </Listing>
@@ -82,7 +82,7 @@ a pair of futures. To begin, each future only hands control back to the runtime
 If you run this, you will see this output:
 
 ```console
-{{#include ../listings/ch17-async-await/listing-17-26/output.txt}}
+{{#include ../listings/ch17-async-await/listing-17-28/output.txt}}
 ```
 
 As with our earlier example, `race` still finishes when `a` finishes. There is
@@ -97,12 +97,12 @@ something we can await!
 However, we can also see the handoff happening in this very example: if we
 removed the `trpl::sleep` at the end of the `a` future, it would complete
 without the `b` future running *at all*. Given that, maybe we could use the
-`sleep` function as a starting point, as in Listing 17-27.
+`sleep` function as a starting point, as in Listing 17-29.
 
-<Listing number="17-27" caption="Using `sleep` to let operations switch off making progress" file-name="src/main.rs">
+<Listing number="17-29" caption="Using `sleep` to let operations switch off making progress" file-name="src/main.rs">
 
 ```rust
-{{#rustdoc_include ../listings/ch17-async-await/listing-17-27/src/main.rs:here}}
+{{#rustdoc_include ../listings/ch17-async-await/listing-17-29/src/main.rs:here}}
 ```
 
 </Listing>
@@ -140,10 +140,10 @@ runtime, using a function named `yield_now`. It does just what it says: hands
 control back to the runtime, so that the runtime can check whether any other
 tasks are ready to make progress.
 
-<Listing number="17-28" caption="Using `yield_now` to let operations switch off making progress" file-name="src/main.rs">
+<Listing number="17-30" caption="Using `yield_now` to let operations switch off making progress" file-name="src/main.rs">
 
 ```rust
-{{#rustdoc_include ../listings/ch17-async-await/listing-17-28/src/main.rs:here}}
+{{#rustdoc_include ../listings/ch17-async-await/listing-17-30/src/main.rs:here}}
 ```
 
 </Listing>
@@ -156,18 +156,18 @@ example, will always sleep for at least a millisecond, even if we pass it a
 lot in one millisecond!
 
 You can see this for yourself by setting up a little benchmark, like the one in
-Listing 17-29. (This is not an especially rigorous way to do performance
+Listing 17-31. (This is not an especially rigorous way to do performance
 testing, but it suffices to show the difference here.) Here, we skip all the
 status printing, pass a one-nanosecond `Duration` to `sleep`, let each future
 run by itself so that they do not interfere with each other, and get rid of all
 the status printing that we did to see the back-and-forth between tasks in
-Listings 17-27 and 17-28. Then we run for 1,000 iterations and see how long
+Listings 17-29 and 17-30. Then we run for 1,000 iterations and see how long
 `sleep` takes vs. `yield_now`.
 
-<Listing number="17-29" caption="Comparing the performance of `sleep` and `yield_now`" file-name="src/main.rs">
+<Listing number="17-31" caption="Comparing the performance of `sleep` and `yield_now`" file-name="src/main.rs">
 
 ```rust
-{{#rustdoc_include ../listings/ch17-async-await/listing-17-29/src/main.rs:here}}
+{{#rustdoc_include ../listings/ch17-async-await/listing-17-31/src/main.rs:here}}
 ```
 
 </Listing>
